Move ssh and mosh into commcare-cloud
