+++ commcare-cloud/commcare_cloud/paths.py
@@ -1,0 +1,4 @@
+import os
+
+REPO_BASE = os.path.expanduser('~/.commcare-cloud/repo')
+ENVIRONMENTS_DIR = os.path.join(REPO_BASE, 'environments')
