+++ scripts/inventory/ssh
@@ -1,13 +1,5 @@
 #!/usr/bin/env bash
-# usage: scripts/inventory/ssh <environment> <group>[:<n>]
-
-SCRIPTS_INVENTORY_DIR=$(dirname $0)
-
-CHOICE=$(SCRIPT=$0 python ${SCRIPTS_INVENTORY_DIR}/getinventory.py $1 $2 | sed 's/:\([0-9]*\)/ -p \1/')
-if [[ -n "$CHOICE" ]]; then
-    shift; shift  # drop environment and group
-    echo "ssh ${CHOICE} $@"
-    ssh ${CHOICE} "$@"
-else
-    exit 1
-fi
+# DEPRECATED use commcare-cloud <env> ssh [<user>@]<group>[:<n>]
+_env=$1
+shift  # drop environment arg
+commcare-cloud $_env ssh "$@"
